feat: fix typescript issues
