+++ composables/libris.ts
@@ -1,4 +1,4 @@
 import { db } from '../plugins/firebase'
 
-export const Libris = async (uid: string | string[], incrementBy: number) =>
-  (await db.ref(`users/${uid}/libris`)).transaction((v) => v + incrementBy)
+export const Libris = (uid: string | string[], incrementBy: number) =>
+  db.ref(`users/${uid}/libris`).transaction(v => v + incrementBy)
