+++ composables/useAuth.ts
@@ -1,10 +1,10 @@
 import { onMounted } from 'vue'
 
-export const useAuth = (mounted) => {
+export const useAuth = (mounted: Function) => {
   const { $auth } = useNuxtApp()
 
   onMounted(() => {
-    $auth.onAuthStateChanged((user) => {
+    $auth.onAuthStateChanged((user: object) => {
       if (!user) return
 
       mounted(user)
