+++ plugins/firebase.ts
@@ -16,11 +16,11 @@ try {
     appId: '1:986081147205:web:fadaee26490f1f173d662c',
     measurementId: 'G-F7Z7BLCQDQ'
   })
-  getPerformance(app)
 } catch (err) {
   firebase.app()
-  getPerformance(app)
 }
+
+getPerformance(app)
 
 const db = firebase.database()
 const auth = firebase.auth()
