+++ plugins/vuetify.ts
@@ -1,21 +1,12 @@
 import { createVuetify } from 'vuetify'
 import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
-import { md3 } from 'vuetify/blueprints'
 import '../assets/scss/vuetify.scss'
 
-export default defineNuxtPlugin((nuxtApp) => {
+export default defineNuxtPlugin(nuxtApp => {
   const vuetify = createVuetify({
     components,
     directives,
-    blueprints: md3,
-    styles: {
-      configFile: 'assets/scss/vuetify.scss'
-    },
-    treeShake: true,
-    options: {
-      customProperties: true
-    },
     theme: {
       defaultTheme: 'dark',
       themes: {
